github: add Release workflow
This workflow builds the extension for various OS/Platform combinations
and publishes the build artifacts as Assets of a new Release.

To trigger a new Release the release commit must be tagged with a tag of
the form `vX.Y.Z`. It is also possible to create a pre-release using a
tag of the form `vX.Y.Z-rcN`.

Once assets are published a user can install the extension using the
following command:

```
gh extensions install hushsecurity/gh-revision
```