gh-pr-revision: new name
This commit renames the tool from `gh-revision` to `gh-pr-revision`.
This is done because `pr` is a more natural namespace for the object
that this tool introduces. Ideally we would like to have a CLI command
`gh pr revision`. However, given that extensions cannot be put under a
builtin command, we rename the tool to have something closer to the
optimum.

The repository will be renamed accordingly as well.