github: verify pull requests do not have merge commits
This is a preparation step for allowing merge commits into the master
branch. Merges of pull requests will be allowed under 2 conditions:

(a) a PR branch doesn't contain merge commits of its own
(b) a PR branch is rebased over master tip

Currently GitHub allows enforcement of (b) in branch protection rules [1].

Enforcement of (a) isn't currently automated by GitHub. Therefore, we
will enforce PR linear history as a build step in Build workflow, and we
will require the Build workflow to succeed to allow the PR merge.

This commit enforces PR linear history as a build step in Build
workflow. The rest is taken care of in repository settings.

[1] https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/managing-rulesets/available-rules-for-rulesets#require-status-checks-to-pass-before-merging