create: store reviewers set in revision metadata
Store reviewers set in revision metadata so that reviewers are never
"forgotten" in CC list.

New revision includes the set of reviewers from all past revisions plus
the ones currently appearing in pull request in GH Api.

Note, it is not possible to (automatically) remove a reviewer once it
appears in any revision metadata.