common: show `gh pr view` error message
`gh pr view` dumps some messages to stdout on error.

Improve our error message to include stdout too if it is available.

The issue originates in failure to authenticate when `gh` CLI is
installed via snap. The plugin failed to run `gh pr view`, and,
before this commit, the error message was hidden.